Use audeer/core/path.py
