+++ audeer/__init__.py
@@ -11,12 +11,12 @@ from audeer.core.io import (
     list_file_names,
     mkdir,
     move_file,
-    path,
     replace_file_extension,
     rmdir,
     safe_path,
     touch,
 )
+from audeer.core.path import path
 from audeer.core.tqdm import (
     format_display_message,
     progress_bar,
